refactor: AppBar 리팩토링
